Move getTestSorterUID back its natural habitat
